Remove alignment in indentation engine
The main reason for this change is to change indentation from

  foobar('arg1', function ()
           print('foobar')
                 end)

to

  foobar('arg1', function ()
      print('foobar')
  end)

The indentation engine is very convoluted, hence the major changes.

The new indentation algorithm is as follows:
Current line indentation is:
+ indentation changes induced by unmatched closers on current line
+ indentation changes induced by previous line
+ previous line indentation
- indent-level if previous line was a continuation
- indent-level

Implementation changes:
* Keyword 'function' is in the same group as '{', '[' and '('.
* Openers yield a simple indent, no alignment.
* Closers at the beginning of the line do not yield any indent for following line.
* Continuing lines are not indented together, only previous line is used for indentation.

As a result, the engine never scans back more than the previous line, except to
check for matching openers. This should result in a sensible performance boost.